requirements.txt
Django
Djangorestframework
psycopg2
